Security & reliability audit fixes across backend, mobile, and web
Backend:
- Wallet: add row locking (SELECT FOR UPDATE), idempotency checks, XAF rounding, refund error handling
- Payments: webhook amount+fee validation, signature 401 response, wallet lock on credit
- Auth: Redis fail-closed (503), email normalization, failed login logging, 256-bit JWT secret
- KYC: DOB/expiry validators, residence_permit support, pending email, rate-limited status check
- Transactions: DB query limit (limit*3), AccountPE deduplication, fetch error counting
- Models: new transaction types (TOPUP_REFUND, WITHDRAWAL_FAILED, CARD_TOPUP), notification TTL
- Schemas: currency whitelist, minimum amount $1, GOLD tier card validation
- Email: SMTP retry with exponential backoff, send result verification
- Notifications: atomic mark-as-read (UPDATE WHERE is_read=False)
- Exchange rate: bounds checking, cache timestamp verification
- Security headers: CSP + Permissions-Policy

Mobile (Flutter):
- Fix Decimal parsing (double.tryParse instead of int.tryParse)
- Card expiry: last day of month calculation
- Simplified isSuccess/isKycVerified/isExpired getters
- Unique notification IDs (microsecondsSinceEpoch)
- Card number auto-hide warning (25s SnackBar before 30s hide)

Web:
- E-nkap credentials moved to environment variables
- Payment callback: default to "failed", added "pending" status handling

Infra:
- Redis healthcheck: --no-auth-warning flag

+++ mobile/lib/screens/wallet/wallet_topup_screen.dart
@@ -434,8 +434,8 @@ class _WalletTopupScreenState extends State<WalletTopupScreen> {
             onChanged: (val) {
               setState(() {
                 if (_isUsdMode) {
-                  final parsed = int.tryParse(val) ?? 0;
-                  _selectedAmountIndex = _usdAmounts.indexOf(parsed);
+                  final parsed = double.tryParse(val) ?? 0.0;
+                  _selectedAmountIndex = _usdAmounts.indexOf(parsed.toInt());
                 } else {
                   _selectedAmountIndex = -1;
                 }
